ai_dm_voice_app/requirements.txt
flask
openai
python-dotenv
discord.py
requests
